+++ dll/src/ClientHandlers/HandleSendChatMessage.cpp
@@ -1,6 +1,8 @@
 #include <unordered_map>
 #include <unordered_set>
 #include <algorithm>
+#include <sstream>
+#include <iomanip>
 #include <regex>
 #include <cwctype>
 #include <filesystem>
@@ -11,16 +13,6 @@
 #include "JSONObject.h"
 #include "URI.h"
 #include "Log.h"
-
-bool HandleChatHelpCommand(std::wstring& name, std::wstring& message, uint32_t& channel, uint8_t& type)
-{
-    GameAPI::SendChatMessage(L"/g, /global[CHANNEL] [ON/OFF]", L"Sends a message to the current global chat channel. If no arguments are specified, displays the current global chat channel.\n    [CHANNEL] - Switches the current global chat channel.\n    [ON/OFF] - Enables or disables global chat.\n ", 0);
-    GameAPI::SendChatMessage(L"/tr, /trade[CHANNEL] [ON/OFF]", L"Sends a message to the current trade chat channel. If no arguments are specified, displays the current trade chat channel.\n    [CHANNEL] - Switches the current trade chat channel.\n    [ON/OFF] - Enables or disables trade chat.\n ", 0);
-    GameAPI::SendChatMessage(L"/c, /challenges[CATEGORY]", L"Displays the user's current challenge progress in the season. If no arguments are specified, displays an overview of all challenge categories.\n    [CATEGORY] - Displays challenges for the specified category.\n ", 0);
-    GameAPI::SendChatMessage(L"/o, /online", L"Displays the number of current users online.\n ", 0);
-    GameAPI::SendChatMessage(L"/h, /help", L"Displays this help message. ", 0);
-    return false;
-}
 
 const std::unordered_map<std::string, uint32_t> challengeCategoryMap =
 {
@@ -33,6 +25,46 @@ const std::unordered_map<std::string, uint32_t> challengeCategoryMap =
     { "Campaign/Story", 7 },
 };
 
+const std::unordered_map<std::wstring, EngineAPI::Color> chatColorMap =
+{
+    { L"a", EngineAPI::Color::AQUA },       { L"aqua", EngineAPI::Color::AQUA },
+    { L"b", EngineAPI::Color::BLUE },       { L"blue", EngineAPI::Color::BLUE },
+    { L"c", EngineAPI::Color::CYAN },       { L"cyan", EngineAPI::Color::CYAN },
+    { L"d", EngineAPI::Color::DARK_GRAY },  { L"darkgray", EngineAPI::Color::DARK_GRAY }, { L"darkgrey", EngineAPI::Color::DARK_GRAY },
+    { L"e", EngineAPI::Color::OLIVE },      { L"olive", EngineAPI::Color::OLIVE },
+    { L"f", EngineAPI::Color::FUSHIA },     { L"fushia", EngineAPI::Color::FUSHIA },
+    { L"g", EngineAPI::Color::GREEN },      { L"green", EngineAPI::Color::GREEN },
+    { L"h", EngineAPI::Color::KHAKI },      { L"khaki", EngineAPI::Color::KHAKI },
+    { L"i", EngineAPI::Color::INDIGO },     { L"indigo", EngineAPI::Color::INDIGO },
+    { L"j", EngineAPI::Color::SALMON },     { L"salmon", EngineAPI::Color::SALMON },
+    { L"k", EngineAPI::Color::BLACK },      { L"black", EngineAPI::Color::BLACK },
+    { L"l", EngineAPI::Color::LIME },       { L"lime", EngineAPI::Color::LIME },
+    { L"m", EngineAPI::Color::MAROON },     { L"maroon", EngineAPI::Color::MAROON },
+    { L"n", EngineAPI::Color::NAVY },       { L"navy", EngineAPI::Color::NAVY },
+    { L"o", EngineAPI::Color::ORANGE },     { L"orange", EngineAPI::Color::ORANGE },
+    { L"p", EngineAPI::Color::PURPLE },     { L"purple", EngineAPI::Color::PURPLE },
+    { L"q", EngineAPI::Color::PINK },       { L"pink", EngineAPI::Color::PINK },
+    { L"r", EngineAPI::Color::RED },        { L"red", EngineAPI::Color::RED },
+    { L"s", EngineAPI::Color::SILVER },     { L"silver", EngineAPI::Color::SILVER },
+    { L"t", EngineAPI::Color::TEAL },       { L"teal", EngineAPI::Color::TEAL },
+    { L"u", EngineAPI::Color::CORNFLOWER }, { L"cornflower", EngineAPI::Color::CORNFLOWER },
+    { L"v", EngineAPI::Color::VIOLET },     { L"violet", EngineAPI::Color::VIOLET },
+    { L"w", EngineAPI::Color::WHITE },      { L"white", EngineAPI::Color::WHITE },
+    { L"x", EngineAPI::Color::GRAY },       { L"gray", EngineAPI::Color::GRAY }, { L"grey", EngineAPI::Color::GRAY },
+    { L"y", EngineAPI::Color::YELLOW },     { L"yellow", EngineAPI::Color::YELLOW },
+    { L"z", EngineAPI::Color::SLATE },      { L"slate", EngineAPI::Color::SLATE },
+};
+
+bool HandleChatHelpCommand(std::wstring& name, std::wstring& message, uint32_t& channel, uint8_t& type)
+{
+    GameAPI::SendChatMessage(L"/g, /global[CHANNEL] [ON/OFF] [COLOR]", L"Sends a message to the current global chat channel. If no arguments are specified, displays the current global chat channel.\n    [CHANNEL] - Switches the current global chat channel.\n    [ON/OFF] - Enables or disables global chat.\n    [COLOR] - Sets the color of global chat to a color alias or a 6-digit hex code.\n ", 0);
+    GameAPI::SendChatMessage(L"/t, /trade[CHANNEL] [ON/OFF] [COLOR]", L"Sends a message to the current trade chat channel. If no arguments are specified, displays the current trade chat channel.\n    [CHANNEL] - Switches the current trade chat channel.\n    [ON/OFF] - Enables or disables trade chat.\n    [COLOR] - Sets the color of trade chat to a color alias or a 6-digit hex code.\n ", 0);
+    GameAPI::SendChatMessage(L"/c, /challenges[CATEGORY]", L"Displays the user's current challenge progress in the season. If no arguments are specified, displays an overview of all challenge categories.\n    [CATEGORY] - Displays challenges for the specified category.\n ", 0);
+    GameAPI::SendChatMessage(L"/o, /online", L"Displays the number of current users online.\n ", 0);
+    GameAPI::SendChatMessage(L"/h, /help", L"Displays this help message. ", 0);
+    return false;
+}
+
 bool HandleChatGlobalCommand(std::wstring& name, std::wstring& message, uint32_t& channel, uint8_t& type)
 {
     ChatClient& chatClient = ChatClient::GetInstance();
@@ -41,9 +73,11 @@ bool HandleChatGlobalCommand(std::wstring& name, std::wstring& message, uint32_t
 
     type = EngineAPI::UI::CHAT_TYPE_GLOBAL;
 
-    std::wstring arg = message;
-    std::transform(arg.begin(), arg.end(), arg.begin(), std::towlower);
-    if (arg == L"on")
+    std::wstring subcommand = message.substr(0, message.find(L" "));
+    std::wstring args = (subcommand.size() == message.size()) ? L"" : message.substr(message.find(L" ") + 1);
+    std::transform(subcommand.begin(), subcommand.end(), subcommand.begin(), std::towlower);
+    std::transform(args.begin(), args.end(), args.begin(), std::towlower);
+    if ((subcommand == L"on") && (subcommand.size() == message.size()))
     {
         if (channel == 0)
             channel = 1;
@@ -51,12 +85,53 @@ bool HandleChatGlobalCommand(std::wstring& name, std::wstring& message, uint32_t
         chatClient.SetCurrentChatChannel(EngineAPI::UI::CHAT_TYPE_GLOBAL, channel);
         return false;
     }
-    else if (arg == L"off")
+    else if ((subcommand == L"off") && (subcommand.size() == message.size()))
     {
         chatClient.SetCurrentChatChannel(EngineAPI::UI::CHAT_TYPE_GLOBAL, 0);
         return false;
     }
-    else if (channel != 0)
+    else if ((subcommand == L"color") || (subcommand == L"colour"))
+    {
+        std::wsmatch match;
+        std::wregex colorRegex(L"^#?([A-Fa-f0-9]{6})$");
+
+        uint32_t colorCode = 0;
+
+        auto it = chatColorMap.find(args);
+        if (it != chatColorMap.end())
+        {
+            EngineAPI::Color color = it->second;
+            colorCode |= (uint32_t)(color._r * 255);
+            colorCode |= ((uint32_t)(color._g * 255) << 8);
+            colorCode |= ((uint32_t)(color._b * 255) << 16);
+            colorCode |= ((uint32_t)(color._a * 255) << 24);
+        }
+        else if (std::regex_match(args, match, colorRegex))
+        {
+            std::wstringstream inputStream;
+            inputStream << std::hex << match.str(1);
+            inputStream >> colorCode;
+
+            // Swap red and blue values to match color code format
+            colorCode |= (colorCode & 0x000000FF) << 24;
+            colorCode = (colorCode & 0xFFFFFF00) | ((colorCode | 0x00FF0000) >> 16);
+            colorCode = (colorCode & 0xFF00FFFF) | ((colorCode & 0xFF000000) >> 8);
+        }
+
+        if (colorCode != 0)
+        {
+            chatWindow.SetChatColor(EngineAPI::UI::CHAT_TYPE_GLOBAL, colorCode);
+
+            std::wstringstream outputStream;
+            outputStream << std::hex << std::uppercase << std::setfill(L'0') << std::setw(2) << (colorCode & 0x0000FF) << std::setw(2) << ((colorCode & 0x00FF00) >> 8) << std::setw(2) << ((colorCode & 0xFF0000) >> 16);
+
+            name = L"Server";
+            message = L"Changed global chat text color to #" + outputStream.str() + L".";
+            return true;
+        }
+    }
+    
+    if (channel != 0)
     {
         if (channel > EngineAPI::UI::CHAT_CHANNEL_MAX)
         {
@@ -108,13 +183,15 @@ bool HandleChatTradeCommand(std::wstring& name, std::wstring& message, uint32_t&
 {
     ChatClient& chatClient = ChatClient::GetInstance();
     EngineAPI::UI::ChatWindow& chatWindow = EngineAPI::UI::ChatWindow::GetInstance();
-    chatWindow.SetChatPrefix(L"/tr ");
+    chatWindow.SetChatPrefix(L"/t ");
 
     type = EngineAPI::UI::CHAT_TYPE_TRADE;
 
-    std::wstring arg = message;
-    std::transform(arg.begin(), arg.end(), arg.begin(), std::towlower);
-    if (arg == L"on")
+    std::wstring subcommand = message.substr(0, message.find(L" "));
+    std::wstring args = (subcommand.size() == message.size()) ? L"" : message.substr(message.find(L" ") + 1);
+    std::transform(subcommand.begin(), subcommand.end(), subcommand.begin(), std::towlower);
+    std::transform(args.begin(), args.end(), args.begin(), std::towlower);
+    if ((subcommand == L"on") && (subcommand.size() == message.size()))
     {
         if (channel == 0)
             channel = 1;
@@ -122,12 +199,53 @@ bool HandleChatTradeCommand(std::wstring& name, std::wstring& message, uint32_t&
         chatClient.SetCurrentChatChannel(EngineAPI::UI::CHAT_TYPE_TRADE, channel);
         return false;
     }
-    else if (arg == L"off")
+    else if ((subcommand == L"off") && (subcommand.size() == message.size()))
     {
         chatClient.SetCurrentChatChannel(EngineAPI::UI::CHAT_TYPE_TRADE, 0);
         return false;
     }
-    else if (channel != 0)
+    else if ((subcommand == L"color") || (subcommand == L"colour"))
+    {
+        std::wsmatch match;
+        std::wregex colorRegex(L"^#?([A-Fa-f0-9]{6})$");
+
+        uint32_t colorCode = 0;
+
+        auto it = chatColorMap.find(args);
+        if (it != chatColorMap.end())
+        {
+            EngineAPI::Color color = it->second;
+            colorCode |= (uint32_t)(color._r * 255);
+            colorCode |= ((uint32_t)(color._g * 255) << 8);
+            colorCode |= ((uint32_t)(color._b * 255) << 16);
+            colorCode |= ((uint32_t)(color._a * 255) << 24);
+        }
+        else if (std::regex_match(args, match, colorRegex))
+        {
+            std::wstringstream inputStream;
+            inputStream << std::hex << match.str(1);
+            inputStream >> colorCode;
+
+            // Swap red and blue values to match color code format
+            colorCode |= (colorCode & 0x000000FF) << 24;
+            colorCode = (colorCode & 0xFFFFFF00) | ((colorCode | 0x00FF0000) >> 16);
+            colorCode = (colorCode & 0xFF00FFFF) | ((colorCode & 0xFF000000) >> 8);
+        }
+
+        if (colorCode != 0)
+        {
+            chatWindow.SetChatColor(EngineAPI::UI::CHAT_TYPE_TRADE, colorCode);
+
+            std::wstringstream outputStream;
+            outputStream << std::hex << std::uppercase << std::setfill(L'0') << std::setw(2) << (colorCode & 0x0000FF) <<  std::setw(2) << ((colorCode & 0x00FF00) >> 8) << std::setw(2) << ((colorCode & 0xFF0000) >> 16);
+
+            name = L"Server";
+            message = L"Changed trade chat text color to #" + outputStream.str() + L".";
+            return true;
+        }
+    }
+
+    if (channel != 0)
     {
         if (channel > EngineAPI::UI::CHAT_CHANNEL_MAX)
         {
@@ -294,6 +412,8 @@ bool HandleChatChallengesCommand(std::wstring& name, std::wstring& message, uint
     pplx::create_task([channel]()
     {
         Client& client = Client::GetInstance();
+        EngineAPI::UI::ChatWindow& chatWindow = EngineAPI::UI::ChatWindow::GetInstance();
+
         web::json::value challenges = GetSeasonChallenges();
         std::unordered_set<uint32_t>& challengeIDs = GetCompletedChallengeIDs();
 
@@ -436,11 +556,11 @@ const std::unordered_map<std::wstring, ChatCommandHandler> chatCommandLookup =
     { L"global",     HandleChatGlobalCommand },
     { L"g",          HandleChatGlobalCommand },
     { L"trade",      HandleChatTradeCommand },
-    { L"tr",         HandleChatTradeCommand },
-    { L"o",          HandleChatOnlineCommand },
+    { L"t",          HandleChatTradeCommand },
     { L"online",     HandleChatOnlineCommand },
-    { L"c",          HandleChatChallengesCommand },
+    { L"o",          HandleChatOnlineCommand },
     { L"challenges", HandleChatChallengesCommand },
+    { L"c",          HandleChatChallengesCommand },
 };
 
 void HandleSendChatMessage(void* _this, const std::wstring& name, const std::wstring& message, uint8_t type, std::vector<uint32_t> targets, uint32_t unk1)